+++ CodeConverter/Util/TypeExtensions.cs
@@ -66,6 +66,9 @@ namespace ICSharpCode.CodeConverter.Util
         /// <param name="symbol">Symbol.</param>
         public static string GetFullMetadataName(this ITypeSymbol symbol)
         {
+            if (symbol is IArrayTypeSymbol ats) {
+                return GetFullMetadataName(ats.ElementType) + "[" + Enumerable.Repeat(',', ats.Rank - 1) + "]";
+            }
             //This is for comaptibility with NR5 reflection name in case of generic types like T1, T2...
             var namedTypeSymbol = symbol as INamedTypeSymbol;
             return namedTypeSymbol != null ? GetFullMetadataName(namedTypeSymbol) : symbol.MetadataName;